requirements.txt
# Enhanced SAM AI Assistant Dependencies
# Modern voice-controlled AI assistant with smooth animations

# Core UI Framework
customtkinter>=5.2.0

# Image Processing
Pillow>=10.0.0

# Computer Vision
opencv-python>=4.8.0

# 3D Visualization
matplotlib>=3.7.0
numpy>=1.24.0
trimesh>=3.20.0
numpy-stl>=2.16.0

# Voice Recognition and TTS
SpeechRecognition>=3.10.0
pyttsx3>=2.90

# System Monitoring
psutil>=5.9.0

# Screen Capture
pyautogui>=0.9.54

# HTTP Requests
requests>=2.31.0

# Audio Processing
pygame>=2.5.0
pycaw>=20230407
comtypes>=1.1.14

# Web Scraping (optional)
beautifulsoup4>=4.12.0

# Search API (optional)
google-search-results>=2.4.0

# Additional Utilities
wikipedia>=1.4.0

# Build Tools (for creating executable)
pyinstaller>=5.13.0

# LLM Providers (optional)
google-generativeai>=0.7.0